+++ R/cbbinom.R
@@ -80,13 +80,15 @@ NULL
 #' @export
 #' @rdname cbbinom
 dcbbinom <- function(x, size, alpha = 1, beta = 1, ncp = 0,
-                     log = FALSE, tol = 1e-6, max_iter = 10000L) {
-  lp <- dcbblp(x, size, alpha, beta, tol, max_iter)
-  # To keep numerical resolution, compute log(p) originally
-  p <- base::log((lp[, 1L, drop = TRUE] - lp[, 2L, drop = TRUE]) / lp[, 3L, drop = TRUE]) +
-    pcbbinom(q = x, size = size, alpha = alpha, beta = beta,
-             lower.tail = TRUE, log.p = TRUE,
-             tol = tol, max_iter = max_iter)
+                     log = FALSE, tol = NULL, max_iter = 10000L) {
+  p <- cpp_dcbbinom(x = as.numeric(x - ncp),
+                    size = as.numeric(size),
+                    alpha = as.numeric(alpha),
+                    beta = as.numeric(beta),
+                    log = TRUE,
+                    tol = if (is.null(tol) == TRUE) NULL else as.numeric(tol),
+                    max_iter = as.integer(max_iter))
+  p[is.na(p) == TRUE] <- -Inf
   if (log == FALSE) {
     p <- exp(p)
   }
@@ -98,14 +100,14 @@ dcbbinom <- function(x, size, alpha = 1, beta = 1, ncp = 0,
 #' @rdname cbbinom
 pcbbinom <- function(q, size, alpha = 1, beta = 1, ncp = 0,
                      lower.tail = TRUE, log.p = FALSE,
-                     tol = 1e-6, max_iter = 10000L) {
+                     tol = NULL, max_iter = 10000L) {
   cpp_pcbbinom(q = as.numeric(q - ncp),
                size = as.numeric(size),
                alpha = as.numeric(alpha),
                beta = as.numeric(beta),
                lower_tail = as.logical(lower.tail[[1L]]),
                log_p = as.logical(log.p[[1L]]),
-               tol = as.numeric(tol),
+               tol = if (is.null(tol) == TRUE) NULL else as.numeric(tol),
                max_iter = as.integer(max_iter))
 }
 
@@ -114,7 +116,7 @@ pcbbinom <- function(q, size, alpha = 1, beta = 1, ncp = 0,
 #' @rdname cbbinom
 qcbbinom <- function(p, size, alpha = 1, beta = 1, ncp = 0,
                      lower.tail = TRUE, log.p = FALSE,
-                     p_tol = 1e-6, p_max_iter = 10000L,
+                     p_tol = NULL, p_max_iter = 10000L,
                      root_tol = 1e-6, root_max_iter = 10000L) {
   cpp_qcbbinom(p = as.numeric(p),
                size = as.numeric(size),
@@ -122,7 +124,7 @@ qcbbinom <- function(p, size, alpha = 1, beta = 1, ncp = 0,
                beta = as.numeric(beta),
                lower_tail = as.logical(lower.tail[[1L]]),
                log_p = as.logical(log.p[[1L]]),
-               p_tol = as.numeric(p_tol),
+               p_tol = if (is.null(p_tol) == TRUE) NULL else as.numeric(p_tol),
                p_max_iter = as.integer(p_max_iter),
                root_tol = as.numeric(root_tol),
                root_max_iter = as.integer(root_max_iter)) + ncp
@@ -132,13 +134,13 @@ qcbbinom <- function(p, size, alpha = 1, beta = 1, ncp = 0,
 #' @export
 #' @rdname cbbinom
 rcbbinom <- function(n, size, alpha = 1, beta = 1, ncp = 0,
-                     p_tol = 1e-6, p_max_iter = 10000L,
+                     p_tol = NULL, p_max_iter = 10000L,
                      root_tol = 1e-6, root_max_iter = 10000L) {
   cpp_rcbbinom(n = as.integer(n[[1L]]),
                size = as.numeric(size),
                alpha = as.numeric(alpha),
                beta = as.numeric(beta),
-               p_tol = as.numeric(p_tol),
+               p_tol = if (is.null(p_tol) == TRUE) NULL else as.numeric(p_tol),
                p_max_iter = as.integer(p_max_iter),
                root_tol = as.numeric(root_tol),
                root_max_iter = as.integer(root_max_iter)) + ncp