+++ src/RcppExports.cpp
@@ -14,7 +14,7 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // cpp_pcbbinom
-NumericVector cpp_pcbbinom(const NumericVector& q, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const NumericVector& tol, const IntegerVector& max_iter);
+NumericVector cpp_pcbbinom(const NumericVector& q, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& tol, const IntegerVector& max_iter);
 static SEXP _cbbinom_cpp_pcbbinom_try(SEXP qSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP lower_tailSEXP, SEXP log_pSEXP, SEXP tolSEXP, SEXP max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -24,7 +24,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const NumericVector& >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< const bool& >::type lower_tail(lower_tailSEXP);
     Rcpp::traits::input_parameter< const bool& >::type log_p(log_pSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type tol(tolSEXP);
+    Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type max_iter(max_iterSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_pcbbinom(q, size, alpha, beta, lower_tail, log_p, tol, max_iter));
     return rcpp_result_gen;
@@ -55,7 +55,7 @@ RcppExport SEXP _cbbinom_cpp_pcbbinom(SEXP qSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // cpp_qcbbinom
-NumericVector cpp_qcbbinom(const NumericVector& p, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const NumericVector& p_tol, const IntegerVector& p_max_iter, const NumericVector& root_tol, const IntegerVector& root_max_iter);
+NumericVector cpp_qcbbinom(const NumericVector& p, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& lower_tail, const bool& log_p, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const NumericVector& root_tol, const IntegerVector& root_max_iter);
 static SEXP _cbbinom_cpp_qcbbinom_try(SEXP pSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP lower_tailSEXP, SEXP log_pSEXP, SEXP p_tolSEXP, SEXP p_max_iterSEXP, SEXP root_tolSEXP, SEXP root_max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -65,7 +65,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const NumericVector& >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< const bool& >::type lower_tail(lower_tailSEXP);
     Rcpp::traits::input_parameter< const bool& >::type log_p(log_pSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type p_tol(p_tolSEXP);
+    Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type p_tol(p_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type p_max_iter(p_max_iterSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type root_tol(root_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type root_max_iter(root_max_iterSEXP);
@@ -97,26 +97,27 @@ RcppExport SEXP _cbbinom_cpp_qcbbinom(SEXP pSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     UNPROTECT(1);
     return rcpp_result_gen;
 }
-// dcbblp
-NumericMatrix dcbblp(const NumericVector& x, const NumericVector& m, const NumericVector& a, const NumericVector& b, const NumericVector& tol, const IntegerVector& max_iter);
-static SEXP _cbbinom_dcbblp_try(SEXP xSEXP, SEXP mSEXP, SEXP aSEXP, SEXP bSEXP, SEXP tolSEXP, SEXP max_iterSEXP) {
+// cpp_dcbbinom
+NumericVector cpp_dcbbinom(const NumericVector& x, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const bool& log, const Nullable<NumericVector>& tol, const IntegerVector& max_iter);
+static SEXP _cbbinom_cpp_dcbbinom_try(SEXP xSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP logSEXP, SEXP tolSEXP, SEXP max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::traits::input_parameter< const NumericVector& >::type x(xSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type m(mSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type a(aSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type b(bSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type tol(tolSEXP);
+    Rcpp::traits::input_parameter< const NumericVector& >::type size(sizeSEXP);
+    Rcpp::traits::input_parameter< const NumericVector& >::type alpha(alphaSEXP);
+    Rcpp::traits::input_parameter< const NumericVector& >::type beta(betaSEXP);
+    Rcpp::traits::input_parameter< const bool& >::type log(logSEXP);
+    Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type max_iter(max_iterSEXP);
-    rcpp_result_gen = Rcpp::wrap(dcbblp(x, m, a, b, tol, max_iter));
+    rcpp_result_gen = Rcpp::wrap(cpp_dcbbinom(x, size, alpha, beta, log, tol, max_iter));
     return rcpp_result_gen;
 END_RCPP_RETURN_ERROR
 }
-RcppExport SEXP _cbbinom_dcbblp(SEXP xSEXP, SEXP mSEXP, SEXP aSEXP, SEXP bSEXP, SEXP tolSEXP, SEXP max_iterSEXP) {
+RcppExport SEXP _cbbinom_cpp_dcbbinom(SEXP xSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP logSEXP, SEXP tolSEXP, SEXP max_iterSEXP) {
     SEXP rcpp_result_gen;
     {
         Rcpp::RNGScope rcpp_rngScope_gen;
-        rcpp_result_gen = PROTECT(_cbbinom_dcbblp_try(xSEXP, mSEXP, aSEXP, bSEXP, tolSEXP, max_iterSEXP));
+        rcpp_result_gen = PROTECT(_cbbinom_cpp_dcbbinom_try(xSEXP, sizeSEXP, alphaSEXP, betaSEXP, logSEXP, tolSEXP, max_iterSEXP));
     }
     Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
     if (rcpp_isInterrupt_gen) {
@@ -137,7 +138,7 @@ RcppExport SEXP _cbbinom_dcbblp(SEXP xSEXP, SEXP mSEXP, SEXP aSEXP, SEXP bSEXP, 
     return rcpp_result_gen;
 }
 // cpp_rcbbinom
-NumericVector cpp_rcbbinom(const int& n, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const NumericVector& p_tol, const IntegerVector& p_max_iter, const NumericVector& root_tol, const IntegerVector& root_max_iter);
+NumericVector cpp_rcbbinom(const int& n, const NumericVector& size, const NumericVector& alpha, const NumericVector& beta, const Nullable<NumericVector>& p_tol, const IntegerVector& p_max_iter, const NumericVector& root_tol, const IntegerVector& root_max_iter);
 static SEXP _cbbinom_cpp_rcbbinom_try(SEXP nSEXP, SEXP sizeSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP p_tolSEXP, SEXP p_max_iterSEXP, SEXP root_tolSEXP, SEXP root_max_iterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -145,7 +146,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const NumericVector& >::type size(sizeSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type alpha(alphaSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type beta(betaSEXP);
-    Rcpp::traits::input_parameter< const NumericVector& >::type p_tol(p_tolSEXP);
+    Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type p_tol(p_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type p_max_iter(p_max_iterSEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type root_tol(root_tolSEXP);
     Rcpp::traits::input_parameter< const IntegerVector& >::type root_max_iter(root_max_iterSEXP);
@@ -178,14 +179,14 @@ RcppExport SEXP _cbbinom_cpp_rcbbinom(SEXP nSEXP, SEXP sizeSEXP, SEXP alphaSEXP,
     return rcpp_result_gen;
 }
 // gen_hypergeo
-double gen_hypergeo(const NumericVector& U, const NumericVector& L, const double& x, const double& tol, const R_xlen_t& max_iter, const bool& check_mode, const bool& log);
+double gen_hypergeo(const NumericVector& U, const NumericVector& L, const double& x, const Nullable<NumericVector>& tol, const R_xlen_t& max_iter, const bool& check_mode, const bool& log);
 static SEXP _cbbinom_gen_hypergeo_try(SEXP USEXP, SEXP LSEXP, SEXP xSEXP, SEXP tolSEXP, SEXP max_iterSEXP, SEXP check_modeSEXP, SEXP logSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::traits::input_parameter< const NumericVector& >::type U(USEXP);
     Rcpp::traits::input_parameter< const NumericVector& >::type L(LSEXP);
     Rcpp::traits::input_parameter< const double& >::type x(xSEXP);
-    Rcpp::traits::input_parameter< const double& >::type tol(tolSEXP);
+    Rcpp::traits::input_parameter< const Nullable<NumericVector>& >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const R_xlen_t& >::type max_iter(max_iterSEXP);
     Rcpp::traits::input_parameter< const bool& >::type check_mode(check_modeSEXP);
     Rcpp::traits::input_parameter< const bool& >::type log(logSEXP);
@@ -222,11 +223,11 @@ RcppExport SEXP _cbbinom_gen_hypergeo(SEXP USEXP, SEXP LSEXP, SEXP xSEXP, SEXP t
 static int _cbbinom_RcppExport_validate(const char* sig) { 
     static std::set<std::string> signatures;
     if (signatures.empty()) {
-        signatures.insert("NumericVector(*cpp_pcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("NumericVector(*cpp_qcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const NumericVector&,const IntegerVector&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("NumericMatrix(*dcbblp)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("NumericVector(*cpp_rcbbinom)(const int&,const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const IntegerVector&,const NumericVector&,const IntegerVector&)");
-        signatures.insert("double(*gen_hypergeo)(const NumericVector&,const NumericVector&,const double&,const double&,const R_xlen_t&,const bool&,const bool&)");
+        signatures.insert("NumericVector(*cpp_pcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&)");
+        signatures.insert("NumericVector(*cpp_qcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const bool&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&,const IntegerVector&)");
+        signatures.insert("NumericVector(*cpp_dcbbinom)(const NumericVector&,const NumericVector&,const NumericVector&,const NumericVector&,const bool&,const Nullable<NumericVector>&,const IntegerVector&)");
+        signatures.insert("NumericVector(*cpp_rcbbinom)(const int&,const NumericVector&,const NumericVector&,const NumericVector&,const Nullable<NumericVector>&,const IntegerVector&,const NumericVector&,const IntegerVector&)");
+        signatures.insert("double(*gen_hypergeo)(const NumericVector&,const NumericVector&,const double&,const Nullable<NumericVector>&,const R_xlen_t&,const bool&,const bool&)");
     }
     return signatures.find(sig) != signatures.end();
 }
@@ -235,7 +236,7 @@ static int _cbbinom_RcppExport_validate(const char* sig) {
 RcppExport SEXP _cbbinom_RcppExport_registerCCallable() { 
     R_RegisterCCallable("cbbinom", "_cbbinom_cpp_pcbbinom", (DL_FUNC)_cbbinom_cpp_pcbbinom_try);
     R_RegisterCCallable("cbbinom", "_cbbinom_cpp_qcbbinom", (DL_FUNC)_cbbinom_cpp_qcbbinom_try);
-    R_RegisterCCallable("cbbinom", "_cbbinom_dcbblp", (DL_FUNC)_cbbinom_dcbblp_try);
+    R_RegisterCCallable("cbbinom", "_cbbinom_cpp_dcbbinom", (DL_FUNC)_cbbinom_cpp_dcbbinom_try);
     R_RegisterCCallable("cbbinom", "_cbbinom_cpp_rcbbinom", (DL_FUNC)_cbbinom_cpp_rcbbinom_try);
     R_RegisterCCallable("cbbinom", "_cbbinom_gen_hypergeo", (DL_FUNC)_cbbinom_gen_hypergeo_try);
     R_RegisterCCallable("cbbinom", "_cbbinom_RcppExport_validate", (DL_FUNC)_cbbinom_RcppExport_validate);
@@ -245,7 +246,7 @@ RcppExport SEXP _cbbinom_RcppExport_registerCCallable() {
 static const R_CallMethodDef CallEntries[] = {
     {"_cbbinom_cpp_pcbbinom", (DL_FUNC) &_cbbinom_cpp_pcbbinom, 8},
     {"_cbbinom_cpp_qcbbinom", (DL_FUNC) &_cbbinom_cpp_qcbbinom, 10},
-    {"_cbbinom_dcbblp", (DL_FUNC) &_cbbinom_dcbblp, 6},
+    {"_cbbinom_cpp_dcbbinom", (DL_FUNC) &_cbbinom_cpp_dcbbinom, 7},
     {"_cbbinom_cpp_rcbbinom", (DL_FUNC) &_cbbinom_cpp_rcbbinom, 8},
     {"_cbbinom_gen_hypergeo", (DL_FUNC) &_cbbinom_gen_hypergeo, 7},
     {"_cbbinom_RcppExport_registerCCallable", (DL_FUNC) &_cbbinom_RcppExport_registerCCallable, 0},