+++ src/hypergeo.h
@@ -7,7 +7,7 @@ using namespace Rcpp;
 double gen_hypergeo(const NumericVector& U,
                     const NumericVector& L,
                     const double& x,
-                    const double& tol,
+                    const Nullable<NumericVector>& tol,
                     const R_xlen_t& max_iter,
                     const bool& check_mode,
                     const bool& log);